substitute a variable <PKG>_DOCDIR for each PKG in dependent hpp packages.
if PKG is insalled,
   <PKG>_DOCDIR is substituted to <PKG>_PREFIX/share/doc/PKG
if PKG is not installed,
   <PKG>_DOCDIR is substituted to prefix/share/hppDoc/PKG

Note: some PKG might not put their doc in share/doc/PKG.
The correct solution is to export datadir (and datarootdir) in the
pkg-config files of PKG, as well as a variable like e.g. docdir which
contain the documentation location.